requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
pandas==2.0.3
python-dotenv==1.0.0
cohere==5.3.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2